Limit discover connection profile to table level when listing tables … (#85)
* Limit discover connection profile to table level when listing tables in Oracle.

* Added comment explaining hierarchy depth and why it is set to 2.

* Renamed to implicate ,it as integration test.

* Added unit test for list tables and to verify hierarchy depth set to 2.

* Resolved comments.

* Add logging for stream validation error in during pipeline assessment

* Add repositories section to pom to fix the build

(cherry picked from commit 062cfad370603bcb1f269d71eaf8dbae3111f0c3)
